fix(gjc): restore projection CI contracts
