+++ tests/unit/gjc/test_gjc_artifacts.py
